+++ ethers-ext/example/accountKey/AccountUpdateWithLegacy.js
@@ -1,0 +1,37 @@
+// AccountKeyLegacy
+// https://docs.klaytn.foundation/docs/learn/accounts/
+
+const { ethers } = require("ethers");
+
+const { Wallet, TxType, AccountKeyType} = require("@klaytn/ethers-ext");
+
+// Using senderPriv == senderNewPriv to execute this example repeatedly.
+// But you might want to register a different private key.
+const senderAddr = "0x24e8efd18d65bcb6b3ba15a4698c0b0d69d13ff7";
+const senderPriv = "0x4a72b3d09c3d5e28e8652e0111f9c4ce252e8299aad95bb219a38eb0a3f4da49";
+const senderNewPriv = "0x4a72b3d09c3d5e28e8652e0111f9c4ce252e8299aad95bb219a38eb0a3f4da49";
+
+const provider = new ethers.providers.JsonRpcProvider("https://public-en-baobab.klaytn.net");
+const wallet = new Wallet(senderPriv, provider);
+
+async function main() {
+  const pub = ethers.utils.computePublicKey(senderNewPriv, true);
+  console.log("pub", pub);
+
+  const tx = {
+    type: TxType.AccountUpdate,
+    from: senderAddr,
+    key: {
+      type: AccountKeyType.Public,
+      key: pub,
+    }
+  };
+
+  const sentTx = await wallet.sendTransaction(tx);
+  console.log("sentTx", sentTx.hash);
+
+  const receipt = await sentTx.wait();
+  console.log("receipt", receipt);
+}
+
+main().catch(console.error);
